requirements.txt
streamlit
streamlit-image-comparison
Pillow
folium
streamlit-folium
google-generativeai
gspread
oauth2client